python-and-data-science/homework_35/requirements.txt
loguru==0.5.3
numpy==1.21.3
Pillow==8.4.0
click==8.0.3
